Revised makefile greatly. Added unit tests for `tfs/path`. Added helpers for unit tests at `src/tests/util.h`. Fixed compilation error with using `TFS_MIN` and `TFS_MAX`. All usages of `TFS_MIN` and `TFS_MAX` not cast their values to the same type. TFS_DIR_MAX_FILE_NAME_LEN is now defined as a `size_t` via a macro.

+++ src/tfs/inode/dir.c
@@ -118,7 +118,7 @@ TfsInodeDirAddEntryResult tfs_inode_dir_add_entry(TfsInodeDir* dir, TfsInodeIdx 
 	if (empty_idx == (size_t)-1) {
 		// Double the current capacity so we don't allocate often
 		// Note: We allocate at least 4 because `2 + 0 == 0`.
-		size_t new_capacity = TFS_MAX(4, 2 * dir->capacity);
+		size_t new_capacity = TFS_MAX((size_t)4, 2 * dir->capacity);
 
 		// Try to allocate
 		// Note: It's fine even if `dir->entries` is `NULL`
